requirements.txt
pyyaml
tinydb
enocean
paho-mqtt
